Update HTML, CSS, JavaScript for Portfolio
Added and finished my early look and feel for the website it is an endless project, but I am happy where I got it now and I will continue to work on it as I graduate. 
I added the CSS the new imgs and logos, the JS for the button and made sure it readable. This is defiantly upgrade to my older 2023 portfolio project which I linked as a live demo in this one!